Extract initialization of the app to a different package
